+++ scanners/headless/base.js
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 
 const puppeteer = require('puppeteer');
 
@@ -41,26 +40,3 @@ var scan = async function (domain, environment, options, scanner, callback) {
 };
 
 module.exports = {scan: scan}
-
-/********************************************************
-** TODO: move this to its own CLI-specific interface file
-*********************************************************/
-
-// the first argument passed in is the name of the scanner
-const scanner = require("../" + process.argv[2]);
-
-// the second argument passed is JSON-serialized data
-const params = JSON.parse(process.argv[3]);
-
-// When executed, run the scan function
-scan(
-  params.domain, params.environment, params.options,
-  scanner,
-  function(err, data) {
-    if (err) {
-      console.error(err);
-      process.exit(1);
-    }
-    console.log(JSON.stringify(data))
-  }
-);
